refactor: kt 린트 포멧팅 및 패키지 변경
