+++ app/src/main/java/woowacourse/shopping/ui/component/CartSkeletonItem.kt
@@ -1,3 +1,5 @@
+@file:Suppress("FunctionName")
+
 package woowacourse.shopping.ui.component
 
 import androidx.compose.foundation.background
